+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# v 1.13.0 (?)
+# v 2.0.0 (?)
 Changes in this release:
+- Update caching mechanism, don't keep project venv in between test session.
+- Halt environment after each test run, resume it before each test run.
 
 # v 1.12.0 (2023-04-03)
 Changes in this release: